+++ src/main/java/com/bewire/DAL/WalletDAO.java
@@ -6,6 +6,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface WalletDAO extends CrudRepository<Wallet,Long>{
+public interface WalletDAO extends CrudRepository<Wallet,Integer>{
     public List<Wallet> findAllByUserId(String id);
 }
